Added sleep and touch after compilation.
Linux VM in Parallels Desktop was seeing some .el files as newer than
their .elc files which was causing a warning.  The fix sleeps after the
all and update_elpa targets and then touches all of the .elc files.
This ensures that all of the .elc files are newer by at least 5 seconds.